refactor tracer wiring into core
